repo_issues.html: show allocated/total sponsor dollars [feat]
In repo_issues.html, add stat items to the header section that show
the allocated/total sponsor dollars for current combination of:

(1) the currently signed-in user (if any)
(2) the owner of the git repo

This commit is the first small step towards implementing
funding/ranking of issues using GitHub Sponsor dollars rather than
direct payments with PayPal.

Showing the number of unallocated sponsor dollars is a good first step
because it requires us to query the sponsor data using GitHub's
GraphQL API. (Note: GitHub does not provide any REST API for querying
GitHub Sponsors data.) It also requires us to think about the database
modeling a bit (see "Implementation note" below).

Note that the number of sponsor dollars available for allocation is
cumulative over time. For example, if a user has sponsored a developer
for $5/month for the past 6 months, then that user will have a total
of $30 sponsor dollars that are available to allocate to the
developer's issues.

Implementation note:

- I changed the name of the `Donation` model to `SponsorAmount` and
added the `sponsor_user_id` as a foreign key.