import: `middleware.py` (`github_autorefresh_token`) [import][feat][dep]
Implement middleware to auto-refresh the user's GitHub OAuth token.

I learned that `django-allauth` has no built-in functionality for
refreshing OAuth tokens [1], so I implemented my own middleware for
auto-refreshing the tokens based on the example at [2].

Note: This commit adds a new dependency on `requests-oauthlib`, for
doing the token refreshing.

[1]: https://codeberg.org/allauth/django-allauth/issues/420
[2]: https://codeberg.org/allauth/django-allauth/issues/420#issuecomment-2295790