repo_issues.html: new modal design (PayPal -> GitHub Sponsors) [UI][wip]
A new (non-functional) mockup of the donate modal UI, due to the
change from PayPal -> Github Sponsor dollars. The new modal will allow
the user to remove funding from an issue at any time, which wasn't
possible when I was using PayPal payments.

The donate modal is just a mockup and is currently non-functional.
For example, changing the donation amount in the edit box does not yet
auto-update the "New total" amount as it should, and the
"Cancel"/"Update" buttons currently do nothing.